revert #116 article section indicator for now
This change was intended to make it easier to navigate longer documents
but so far we didn't get to implement a proper Table Of Contents view
on top of the work that has already been done in PR #116.

With the upcoming responsiveness improvements the layout constraints
changed and keeping this around would have been significant work.

I'm also slowly getting my head around having a Table of Contents as
a secondary sidebar similar to how Docosaurus and other tools handle
this — Something the existing approach isn't really compatible with.